refactor: Jail module
Reorganized jail system code to a new folder under bot_modules.

Issue: None
